Simplify smart contract (#2)
Simplify smart contract part
