+++ smart-contract/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2021"
 odra = { version = "0.9.0", features = [], default-features = false }
 odra-modules = "0.9.0"
 odra-casper-livenet-env = { version = "0.9.0", optional = true }
-odra_cep47 = { git = "https://github.com/odradev/odra-cep47.git" }
 
 [dev-dependencies]
 odra-test = { version = "0.9.0", features = [], default-features = false }
