+++ smart-contract/bin/build_contract.rs
