+++ smart-contract/bin/build_schema.rs
@@ -1,7 +1,5 @@
 #![doc = "Binary for building schema definitions from odra contracts."]
 #[allow(unused_imports)]
-use lottery;
-
 #[cfg(not(target_arch = "wasm32"))]
 extern "Rust" {
     fn module_schema() -> odra::contract_def::ContractBlueprint;
@@ -16,7 +14,7 @@ fn main() {
     }
 
     let module = std::env::var("ODRA_MODULE").expect("ODRA_MODULE environment variable is not set");
-    let mut schema_file = std::fs::File::create(&format!("resources/{}_schema.json", module))
+    let mut schema_file = std::fs::File::create(format!("resources/{}_schema.json", module))
         .expect("Failed to create schema file");
     let json = schema.as_json().expect("Failed to convert schema to JSON");
     std::io::Write::write_all(&mut schema_file, &json.into_bytes())
