--- smart-contract/faucet/CHANGELOG.md
@@ -1,6 +1,0 @@
-# Changelog
-
-Changelog for `facuet`.
-
-## [0.1.0] - 2024-03-14
-### Added
