--- smart-contract/faucet/Odra.toml
@@ -1,2 +1,0 @@
-[[contracts]]
-fqn = "faucet::Faucet"
