--- smart-contract/faucet/bin/build_contract.rs
@@ -1,5 +1,0 @@
-#![doc = "Binary for building wasm files from odra contracts."]
-#![no_std]
-#![no_main]
-#![allow(unused_imports, clippy::single_component_path_imports)]
-use faucet;
