--- smart-contract/faucet/src/lib.rs
@@ -1,5 +1,0 @@
-#![cfg_attr(not(test), no_std)]
-#![cfg_attr(not(test), no_main)]
-extern crate alloc;
-
-pub mod faucet;
