--- smart-contract/lottery/build.rs
@@ -1,6 +1,0 @@
-//! Odra's contracts build script.
-
-/// Uses the ENV variable `ODRA_MODULE` to set the `odra_module` cfg flag.
-pub fn main() {
-    odra_build::build();
-}
