--- smart-contract/lottery/rust-toolchain
@@ -1,1 +1,0 @@
-nightly-2024-01-26
