+++ smart-contract/src/lib.rs
